split integration controller into pieces - prepare for testing
